+++ app/src/main/java/com/example/smartfarming/data/network/resources/garden_resource/request/Location.kt
@@ -1,5 +1,5 @@
 package com.example.smartfarming.data.network.resources.garden_resource.request
 
-data class Location(val latitude: Int = 0,
+data class Location(val latitude: Double = 0.0,
                     val id: Int = 0,
-                    val longitude: Int = 0)
+                    val longitude: Double = 0.0)
